Docs and better code organization
