+++ src/Tickets/Commerce/Gateways/Square/Syncs/Items_Sync.php
@@ -1,0 +1,370 @@
+<?php
+/**
+ * Syncs tickets with Square controller.
+ *
+ * @since TBD
+ *
+ * @package TEC\Tickets\Commerce\Gateways\Square\Syncs
+ */
+
+namespace TEC\Tickets\Commerce\Gateways\Square\Syncs;
+
+use TEC\Common\Contracts\Provider\Controller as Controller_Contract;
+use WP_Query;
+use TEC\Common\Contracts\Container;
+use TEC\Tickets\Flexible_Tickets\Series_Passes\Series_Passes;
+use Tribe__Tickets__Tickets as Tickets;
+use TEC\Tickets\Commerce\Gateways\Square\Requests;
+use TEC\Tickets\Commerce\Gateways\Square\Syncs\Objects\Item;
+use TEC\Tickets\Ticket_Data;
+use TEC\Tickets\Commerce\Gateways\Square\Syncs\Controller as Sync_Controller;
+
+/**
+ * Class Tickets_Sync
+ *
+ * @since TBD
+ *
+ * @package TEC\Tickets\Commerce\Gateways\Square\Syncs
+ */
+class Items_Sync extends Controller_Contract {
+	/**
+	 * The action that syncs tickets with Square.
+	 *
+	 * @since TBD
+	 *
+	 * @var string
+	 */
+	public const HOOK_INIT_SYNC_ACTION = 'tec_tickets_commerce_square_sync_tickets';
+
+	/**
+	 * The action that syncs an individual event and its tickets with Square.
+	 *
+	 * @since TBD
+	 *
+	 * @var string
+	 */
+	public const HOOK_SYNC_EVENT_ACTION = 'tec_tickets_commerce_square_sync_event';
+
+	/**
+	 * The action that syncs the tickets of a ticket-able post type with Square.
+	 *
+	 * @since TBD
+	 *
+	 * @var string
+	 */
+	public const HOOK_SYNC_ACTION = 'tec_tickets_commerce_square_sync_post_type';
+
+	/**
+	 * The remote objects instance.
+	 *
+	 * @since TBD
+	 *
+	 * @var Remote_Objects
+	 */
+	private Remote_Objects $remote_objects;
+
+	/**
+	 * The ticket data instance.
+	 *
+	 * @since TBD
+	 *
+	 * @var Ticket_Data
+	 */
+	private Ticket_Data $ticket_data;
+
+	/**
+	 * Constructor.
+	 *
+	 * @since TBD
+	 *
+	 * @param Container      $container      The container instance.
+	 * @param Remote_Objects $remote_objects The remote objects instance.
+	 * @param Ticket_Data    $ticket_data    The ticket data instance.
+	 */
+	public function __construct( Container $container, Remote_Objects $remote_objects, Ticket_Data $ticket_data ) {
+		parent::__construct( $container );
+		$this->remote_objects = $remote_objects;
+		$this->ticket_data    = $ticket_data;
+	}
+
+	/**
+	 * Register the controller.
+	 *
+	 * @since TBD
+	 *
+	 * @return void
+	 */
+	public function do_register(): void {
+		add_action( self::HOOK_INIT_SYNC_ACTION, [ $this, 'schedule_sync_for_each_post_type' ] );
+		add_action( self::HOOK_SYNC_ACTION, [ $this, 'sync_post_type' ] );
+		add_action( self::HOOK_SYNC_EVENT_ACTION, [ $this, 'sync_event' ] );
+	}
+
+	/**
+	 * Unregister the controller.
+	 *
+	 * @since TBD
+	 *
+	 * @return void
+	 */
+	public function unregister(): void {
+		remove_action( self::HOOK_INIT_SYNC_ACTION, [ $this, 'schedule_sync_for_each_post_type' ] );
+		remove_action( self::HOOK_SYNC_ACTION, [ $this, 'sync_post_type' ] );
+		remove_action( self::HOOK_SYNC_EVENT_ACTION, [ $this, 'sync_event' ] );
+	}
+
+	/**
+	 * Sync the tickets.
+	 *
+	 * @since TBD
+	 *
+	 * @return void
+	 */
+	public function schedule_sync_for_each_post_type(): void {
+		if ( Sync_Controller::is_sync_completed() ) {
+			return;
+		}
+
+		$ticket_able_post_types = (array) tribe_get_option( 'ticket-enabled-post-types', [] );
+
+		tribe_update_option( sprintf( Sync_Controller::OPTION_SYNC_ACTIONS_IN_PROGRESS, 'default' ), time() );
+
+		foreach ( $ticket_able_post_types as $ticket_able_post_type ) {
+			as_unschedule_action( self::HOOK_SYNC_ACTION, [ $ticket_able_post_type ], Sync_Controller::AS_SYNC_ACTION_GROUP );
+			as_schedule_single_action( time(), self::HOOK_SYNC_ACTION, [ $ticket_able_post_type ], Sync_Controller::AS_SYNC_ACTION_GROUP );
+		}
+	}
+
+	/**
+	 * Sync the tickets for a ticket type.
+	 *
+	 * @since TBD
+	 *
+	 * @param string $ticket_able_post_type The ticket-able post type.
+	 *
+	 * @return void
+	 */
+	public function sync_post_type( string $ticket_able_post_type ): void {
+		if ( Sync_Controller::is_sync_completed() ) {
+			return;
+		}
+
+		$args = [
+			/**
+			 * Filters the number of "events" aka ticket-able post types to sync with Square per batch.
+			 *
+			 * Each "event" is created as an Event Item in Square. Each ticket the event is created with is created as a Variation within the Event Item.
+			 *
+			 * We can send up to 1000 items per batch to Square. We can't send less than 1 as well -_- !
+			 *
+			 * @since TBD
+			 *
+			 * @param int $posts_per_page The number of posts to sync.
+			 */
+			'posts_per_page'    => min( 1000, max( 1, (int) apply_filters( 'tec_tickets_commerce_square_sync_post_type_posts_per_page', 100 ) ) ),
+			'post_type'         => $ticket_able_post_type,
+			'tribe-has-tickets' => true,
+			'post_status'       => 'publish',
+			'fields'            => 'ids',
+			'meta_query'        => [
+				[
+					'key'     => Item::SQUARE_SYNCED_META,
+					'compare' => 'NOT EXISTS',
+				],
+			],
+		];
+
+		$query = new WP_Query(
+			/**
+			 * Filters the query arguments for the ticket-able post type tickets sync.
+			 *
+			 * @since TBD
+			 *
+			 * @param array $args The query arguments.
+			 */
+			(array) apply_filters( 'tec_tickets_commerce_square_sync_post_type_query_args', $args )
+		);
+
+		if ( ! $query->have_posts() ) {
+			tribe_remove_option( sprintf( Sync_Controller::OPTION_SYNC_ACTIONS_COMPLETED, 'default' ) );
+			// Post type is synced! Now on to sync the inventory.
+			as_schedule_single_action( time(), Inventory_Sync::HOOK_SYNC_ACTION, [ $ticket_able_post_type ], Sync_Controller::AS_SYNC_ACTION_GROUP );
+			return;
+		}
+
+		tribe_update_option( sprintf( Sync_Controller::OPTION_SYNC_ACTIONS_IN_PROGRESS, $ticket_able_post_type ), time() );
+
+		// Reschedules itself to continue in 2 minutes.
+		as_schedule_single_action( time() + MINUTE_IN_SECONDS * 2, self::HOOK_SYNC_ACTION, [ $ticket_able_post_type ], Sync_Controller::AS_SYNC_ACTION_GROUP );
+
+		$post_ids = $query->posts;
+
+		$batch = [];
+
+		foreach ( $post_ids as $post_id ) {
+			$tickets = $this->sync_event( $post_id, false );
+
+			if ( ! $tickets ) {
+				update_post_meta( $post_id, Item::SQUARE_SYNCED_META, time() );
+				continue;
+			}
+
+			$batch[ $post_id ] = $tickets;
+		}
+
+		$batch = array_filter( $batch );
+
+		if ( empty( $batch ) ) {
+			return;
+		}
+
+		$this->process_batch( $batch );
+	}
+
+	/**
+	 * Sync the event.
+	 *
+	 * @since TBD
+	 *
+	 * @param int $event_id The event ID.
+	 * @param bool $execute Whether to execute the sync.
+	 *
+	 * @return array The tickets.
+	 */
+	public function sync_event( int $event_id, bool $execute = true ): array {
+		$tickets_stats = $this->ticket_data->get_posts_tickets_data( $event_id, [ 'rsvp', Series_Passes::TICKET_TYPE ] );
+
+		if (
+			empty( $tickets_stats['tickets_on_sale'] ) &&
+			empty( $tickets_stats['tickets_about_to_go_to_sale'] ) &&
+			empty( $tickets_stats['tickets_have_ended_sales'] )
+		) {
+			return [];
+		}
+
+		$ticket_ids = array_unique(
+			array_merge(
+				$tickets_stats['tickets_on_sale'],
+				$tickets_stats['tickets_about_to_go_to_sale'],
+				$tickets_stats['tickets_have_ended_sales']
+			)
+		);
+
+		$tickets = array_filter(
+			array_map(
+				static fn ( $ticket_id ) => Tickets::load_ticket_object( $ticket_id ),
+				$ticket_ids
+			)
+		);
+
+		if ( ! $execute ) {
+			return $tickets;
+		}
+
+		$this->process_batch( [ $event_id => $tickets ] );
+
+		return $tickets;
+	}
+
+	/**
+	 * Process the batch.
+	 *
+	 * @since TBD
+	 *
+	 * @param array $batch The batch.
+	 *
+	 * @return void
+	 */
+	protected function process_batch( array $batch ): void {
+		$square_batches = $this->remote_objects->transform_batch( $batch );
+
+		$idempotency_key = uniqid( 'tec-square-', true );
+
+		$args = [
+			'body'    => [
+				'idempotency_key' => $idempotency_key,
+				'batches'         => $square_batches,
+			],
+			'headers' => [
+				'Content-Type' => 'application/json',
+			],
+		];
+
+		$response = Requests::post(
+			'catalog/batch-upsert',
+			[],
+			$args
+		);
+
+		if ( ! empty( $response['errors'] ) ) {
+			do_action( 'tribe_log', 'error', 'Square Sync', (array) $response['errors'] );
+			return;
+		}
+
+		if ( ! empty( $response['id_mappings'] ) ) {
+			foreach ( $response['id_mappings'] as $id_mapping ) {
+				/**
+				 * Fires when a ticket ID mapping is received from Square.
+				 *
+				 * @since TBD
+				 *
+				 * @param string $object_id The object ID.
+				 * @param array  $id_mapping The ID mapping.
+				 */
+				do_action( 'tec_tickets_commerce_square_sync_ticket_id_mapping_' . $id_mapping['client_object_id'], $id_mapping['object_id'], $id_mapping );
+
+				/**
+				 * Fires when a ticket ID mapping is received from Square.
+				 *
+				 * @since TBD
+				 *
+				 * @param array $id_mapping The ID mapping.
+				 */
+				do_action( 'tec_tickets_commerce_square_sync_ticket_id_mapping', $id_mapping );
+			}
+		}
+
+		if ( empty( $response['objects'] ) ) {
+			do_action( 'tribe_log', 'error', 'Square Sync', [ 'idempotency_key' => $idempotency_key, 'response' => $response ] );
+			return;
+		}
+
+		foreach ( $response['objects'] as $object ) {
+			$this->fire_sync_object_hooks( $object );
+		}
+	}
+
+	/**
+	 * Fires the sync object hooks.
+	 *
+	 * @since TBD
+	 *
+	 * @param array $object The object.
+	 *
+	 * @return void
+	 */
+	protected function fire_sync_object_hooks( array $object ): void {
+		/**
+		 * Fires when a object is received from Square.
+		 *
+		 */
+		do_action( 'tec_tickets_commerce_square_sync_object_' . $object['id'], $object );
+
+		/**
+		 * Fires when a object is received from Square.
+		 *
+		 * @since TBD
+		 *
+		 * @param array $object The sync object.
+		 */
+		do_action( 'tec_tickets_commerce_square_sync_object', $object );
+
+		if ( empty( $object['item_data']['variations'] ) || ! is_array( $object['item_data']['variations'] ) ) {
+			return;
+		}
+
+		foreach ( $object['item_data']['variations'] as $variation ) {
+			$this->fire_sync_object_hooks( $variation );
+		}
+	}
+}
